backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["gmarinella9 <giuliano.b.marinella@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.68.1"
sqlmodel = "^0.0.4"
uvicorn = {extras = ["standard"], version = "^0.15.0"}
requests = "^2.26.0"
strawberry-graphql = {extras = ["debug-server"], version = "^0.75.1"}
pytest-cov = "^2.12.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
